CORFU with group-failover experiments

+++ ACE/apps/JAWS/clients/Blobby/Blob_Handler.cpp
@@ -280,7 +280,8 @@ ACE_Blob_Writer::send_request (void)
 
   // Create the header, store the actual length in mesglen.
   mesglen = ACE_OS::sprintf (mesg, "%s /%s %s " ACE_SIZE_T_FORMAT_SPECIFIER_ASCII "\n\n",
-                             request_prefix_, filename_, request_suffix_, length_);
+                             request_prefix_, filename_, request_suffix_,
+                             (unsigned long)length_);
 
   // Send the header followed by the data
